+++ NEWS.md
@@ -15,6 +15,8 @@
 * Added the internal class template `MatrixColumnVisitor` to represent 
   `VectorVisitor` concept for a column that is a `matrix`. Part of #602  
 
+* `arrange` can sort based on `data.frame` columns (#765).   
+  
 # dplyr 0.3.0.9000
 
 * Hybrid version of `lead` and `lag` are enabled, only for the simple 2 args case